requirements.txt
kafka-python==2.1.0
websocket-client==1.6.2
